Lifted sizesInStock state to parent. We filter the list in the parent, set parent's sizesInStock state to the result, then pass sizesInStock state to child. Still doesn't work.
